Fixes and adding instrumentation for closeable http client
