--- instrumentation/apache-httpclient/apache-httpclient-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v4_0/ApacheHttpClientInstrumentation.java
@@ -1,118 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0;
-
-import static io.opentelemetry.javaagent.bootstrap.Java8BytecodeBridge.currentContext;
-import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.extendsClass;
-import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.hasClassesNamed;
-import static io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0.ApacheHttpClientSingletons.createOrGetBytesTransferMetrics;
-import static io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0.ApacheHttpClientSingletons.instrumenter;
-import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
-import static net.bytebuddy.matcher.ElementMatchers.isFinal;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpEntityEnclosingRequest;
-import org.apache.http.HttpHost;
-import org.apache.http.HttpRequest;
-import org.apache.http.HttpResponse;
-import org.apache.http.protocol.HttpContext;
-
-public class ApacheHttpClientInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<ClassLoader> classLoaderOptimization() {
-    return hasClassesNamed("org.apache.http.impl.client.AbstractHttpClient");
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return extendsClass(named("org.apache.http.impl.client.AbstractHttpClient"));
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        isMethod()
-            .and(named("execute"))
-            .and(isPublic())
-            .and(isFinal())
-            .and(not(isAbstract()))
-            .and(takesArguments(3))
-            .and(takesArgument(0, named("org.apache.http.HttpHost")))
-            .and(takesArgument(1, named("org.apache.http.HttpRequest")))
-            .and(takesArgument(2, named("org.apache.http.protocol.HttpContext"))),
-        this.getClass().getName() + "$RequestAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class RequestAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.Argument(0) HttpHost host,
-        @Advice.Argument(1) HttpRequest request,
-        @Advice.Argument(2) HttpContext httpContext,
-        @Advice.Local("otelRequest") ApacheHttpClientRequest otelRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-
-      otelRequest = new ApacheHttpClientRequest(parentContext, host, request);
-      if (request instanceof HttpEntityEnclosingRequest) {
-        HttpEntity entity = ((HttpEntityEnclosingRequest) request).getEntity();
-        if (entity != null) {
-          long contentLength = entity.getContentLength();
-          BytesTransferMetrics metrics = createOrGetBytesTransferMetrics(parentContext);
-          metrics.setRequestContentLength(contentLength);
-          HttpEntity wrappedHttpEntity = new WrappedHttpEntity(parentContext, entity);
-          ((HttpEntityEnclosingRequest) request).setEntity(wrappedHttpEntity);
-        }
-      }
-
-      if (instrumenter().shouldStart(parentContext, otelRequest)) {
-        context = instrumenter().start(parentContext, otelRequest);
-        scope = context.makeCurrent();
-      }
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Return HttpResponse response,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelRequest") ApacheHttpClientRequest otelRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope != null) {
-        scope.close();
-        if (response != null) {
-          HttpEntity entity = response.getEntity();
-          if (entity != null) {
-            long contentLength = entity.getContentLength();
-            Context parentContext = otelRequest.getParentContext();
-            BytesTransferMetrics metrics = createOrGetBytesTransferMetrics(parentContext);
-            metrics.setResponseContentLength(contentLength);
-          }
-        }
-        if (throwable != null) {
-          instrumenter().end(context, otelRequest, null, throwable);
-        } else {
-          instrumenter().end(context, otelRequest, response, null);
-        }
-      }
-    }
-  }
-}
