+++ instrumentation/apache-httpclient/apache-httpclient-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v4_0/ApacheHttpClientInstrumentationModule.java
@@ -5,11 +5,10 @@
 
 package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0;
 
-import static java.util.Collections.singletonList;
-
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import java.util.ArrayList;
 import java.util.List;
 
 @AutoService(InstrumentationModule.class)
@@ -21,6 +20,9 @@ public class ApacheHttpClientInstrumentationModule extends InstrumentationModule
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new ApacheHttpClientInstrumentation());
+    List<TypeInstrumentation> instrumentationList = new ArrayList<>();
+    instrumentationList.add(new ApacheAbstractHttpClientInstrumentation());
+    instrumentationList.add(new ApacheCloseableHttpClientInstrumentation());
+    return instrumentationList;
   }
 }
